GtkFileChooserWidget: Move the action and operation_mode fields to a GtkFileChooserState struct
We introduce a new GtkFileChooserState struct, which we'll use as a
model in the sense of Functional Reactive Programming.  Events/signals
from the file chooser's widgets will just cause things to change in a
goal State.  Eventually, the file chooser will sync its widgets to
this goal State (load a different folder, switch widgets around,
etc.).

Hopefully this will also make it easier to test the individual parts
of the file chooser.